chore: Follow up add chat or subspace UX
