+++ lib/pages/chat_list/chat_list.dart
@@ -105,106 +105,17 @@ class ChatListController extends State<ChatList>
   String? _activeSpaceId;
   String? get activeSpaceId => _activeSpaceId;
 
-  void setActiveSpace(String spaceId) => setState(() {
-        _activeSpaceId = spaceId;
-      });
+  void setActiveSpace(String spaceId) async {
+    await Matrix.of(context).client.getRoomById(spaceId)!.postLoad();
+
+    setState(() {
+      _activeSpaceId = spaceId;
+    });
+  }
+
   void clearActiveSpace() => setState(() {
         _activeSpaceId = null;
       });
-
-  void addChatAction() async {
-    if (activeSpaceId == null) {
-      context.go('/rooms/newprivatechat');
-      return;
-    }
-
-    final roomType = await showConfirmationDialog(
-      context: context,
-      title: L10n.of(context)!.addChatOrSubSpace,
-      actions: [
-        AlertDialogAction(
-          key: AddRoomType.subspace,
-          label: L10n.of(context)!.createNewSpace,
-        ),
-        AlertDialogAction(
-          key: AddRoomType.chat,
-          label: L10n.of(context)!.createGroup,
-        ),
-      ],
-    );
-    if (roomType == null) return;
-
-    final names = await showTextInputDialog(
-      context: context,
-      title: roomType == AddRoomType.subspace
-          ? L10n.of(context)!.createNewSpace
-          : L10n.of(context)!.createGroup,
-      textFields: [
-        DialogTextField(
-          hintText: roomType == AddRoomType.subspace
-              ? L10n.of(context)!.spaceName
-              : L10n.of(context)!.groupName,
-          minLines: 1,
-          maxLines: 1,
-          maxLength: 64,
-          validator: (text) {
-            if (text == null || text.isEmpty) {
-              return L10n.of(context)!.pleaseChoose;
-            }
-            return null;
-          },
-        ),
-        DialogTextField(
-          hintText: L10n.of(context)!.chatDescription,
-          minLines: 4,
-          maxLines: 8,
-          maxLength: 255,
-        ),
-      ],
-      okLabel: L10n.of(context)!.create,
-      cancelLabel: L10n.of(context)!.cancel,
-    );
-    if (names == null) return;
-    final client = Matrix.of(context).client;
-    final result = await showFutureLoadingDialog(
-      context: context,
-      future: () async {
-        late final String roomId;
-        final activeSpace = client.getRoomById(activeSpaceId!)!;
-        await activeSpace.postLoad();
-
-        if (roomType == AddRoomType.subspace) {
-          roomId = await client.createSpace(
-            name: names.first,
-            topic: names.last.isEmpty ? null : names.last,
-            visibility: activeSpace.joinRules == JoinRules.public
-                ? sdk.Visibility.public
-                : sdk.Visibility.private,
-          );
-        } else {
-          roomId = await client.createGroupChat(
-            groupName: names.first,
-            preset: activeSpace.joinRules == JoinRules.public
-                ? CreateRoomPreset.publicChat
-                : CreateRoomPreset.privateChat,
-            visibility: activeSpace.joinRules == JoinRules.public
-                ? sdk.Visibility.public
-                : sdk.Visibility.private,
-            initialState: names.length > 1 && names.last.isNotEmpty
-                ? [
-                    sdk.StateEvent(
-                      type: sdk.EventTypes.RoomTopic,
-                      content: {'topic': names.last},
-                    ),
-                  ]
-                : null,
-          );
-        }
-        await activeSpace.setSpaceChild(roomId);
-      },
-    );
-    if (result.error != null) return;
-  }
 
   void onChatTap(Room room) async {
     if (room.membership == Membership.invite) {
@@ -1003,8 +914,6 @@ enum InviteActions {
   decline,
   block,
 }
-
-enum AddRoomType { chat, subspace }
 
 enum ChatContextAction {
   open,
